refactor : 공통 기능 분리를 위해 interface를 abstract class로 변경
