+++ src/main/java/blackjack/domain/player/Dealer.java
@@ -5,7 +5,7 @@ import blackjack.domain.card.Cards;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Dealer implements Player {
+public class Dealer extends Player {
 
     public static final String DEALER_NAME = "딜러";
 
@@ -13,31 +13,15 @@ public class Dealer implements Player {
     private static final int DEALER_RECEIVE_STANDARD = 16;
 
     private final String name;
-    private final Cards cards;
 
     public Dealer() {
+        super(new Cards());
         name = DEALER_NAME;
-        cards = new Cards();
-    }
-
-    @Override
-    public void receiveCard(final Card card) {
-        cards.save(card);
     }
 
     @Override
     public List<Card> openCards() {
         return new ArrayList<>(cards.getCards().subList(0, DEALER_OPEN_CARD_NUMBER));
-    }
-
-    @Override
-    public List<Card> showCards() {
-        return List.copyOf(cards.getCards());
-    }
-
-    @Override
-    public int calculateResult() {
-        return cards.calculateTotalPoint();
     }
 
     @Override
