+++ src/main/java/blackjack/domain/player/Gamer.java
@@ -7,20 +7,19 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
-public class Gamer implements Player {
+public class Gamer extends Player {
 
     public static final int LIMIT_GAMER_TOTAL_POINT = 21;
 
     private static final int GAMER_OPEN_CARD_NUMBER = 2;
 
     private final String name;
-    private final Cards cards;
 
     public Gamer(final String name) {
+        super(new Cards());
         checkEmptyName(name);
         checkBannedName(name);
         this.name = name;
-        cards = new Cards();
     }
 
     private void checkEmptyName(final String name) {
@@ -36,11 +35,6 @@ public class Gamer implements Player {
     }
 
     @Override
-    public void receiveCard(final Card card) {
-        cards.save(card);
-    }
-
-    @Override
     public List<Card> openCards() {
         return new ArrayList<>(cards.getCards().subList(0, GAMER_OPEN_CARD_NUMBER));
     }
@@ -48,11 +42,6 @@ public class Gamer implements Player {
     @Override
     public List<Card> showCards() {
         return List.copyOf(cards.getCards());
-    }
-
-    @Override
-    public int calculateResult() {
-        return cards.calculateTotalPoint();
     }
 
     @Override
